# web/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "StigVidd",
    "description": "Stigvidd API. This is free software, licensed under the GNU Affero General Public License version 3 or later. Under section 13 of that licence you are entitled to the Corresponding Source of this running service: https://github.com/themalind/stigvidd",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/v1/Facilities": {
      "get": {
        "tags": [
          "Facilities"
        ],
        "operationId": "Facilities_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Facilities"
        ],
        "operationId": "Facilities_Create",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFacilityRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilityResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Facilities/{identifier}": {
      "get": {
        "tags": [
          "Facilities"
        ],
        "operationId": "Facilities_GetByIdentifier",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Facilities"
        ],
        "operationId": "Facilities_Delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Facilities/update/{identifier}": {
      "put": {
        "tags": [
          "Facilities"
        ],
        "operationId": "Facilities_Update",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFacilityRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilityResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Facilities/{identifier}/images": {
      "post": {
        "tags": [
          "Facilities"
        ],
        "operationId": "Facilities_AddFacilityImages",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "images": {
                    "type": "array",
                    "nullable": true,
                    "items": {}
                  },
                  "MaxWidth": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "MaxHeight": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "Quality": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "Format": {
                    "type": "string",
                    "nullable": true
                  },
                  "CropX": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropY": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropWidth": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropHeight": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacilityImageResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Facilities/images/{imageIdentifier}": {
      "delete": {
        "tags": [
          "Facilities"
        ],
        "operationId": "Facilities_DeleteFacilityImage",
        "parameters": [
          {
            "name": "imageIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Friends": {
      "get": {
        "tags": [
          "Friends"
        ],
        "operationId": "Friends_GetFriends",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Friends/requests/incoming": {
      "get": {
        "tags": [
          "Friends"
        ],
        "operationId": "Friends_GetIncomingFriendRequests",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Friends/requests/outgoing": {
      "get": {
        "tags": [
          "Friends"
        ],
        "operationId": "Friends_GetOutgoingFriendRequests",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutgoingFriendRequestResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Friends/requests": {
      "post": {
        "tags": [
          "Friends"
        ],
        "operationId": "Friends_SendFriendRequest",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFriendRequestRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Friends/requests/accept/{requesterIdentifier}": {
      "put": {
        "tags": [
          "Friends"
        ],
        "operationId": "Friends_AcceptFriendRequest",
        "parameters": [
          {
            "name": "requesterIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Friends/reject/{otherIdentifier}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "operationId": "Friends_RemoveConnection",
        "parameters": [
          {
            "name": "otherIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Friends/{friendIdentifier}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "operationId": "Friends_RemoveFriend",
        "parameters": [
          {
            "name": "friendIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FacilityResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "facilityType"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "facilityType": {
            "type": "integer",
            "format": "int32"
          },
          "isAccessible": {
            "type": "boolean"
          },
          "latitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateFacilityRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "facilityType",
          "isAccessible",
          "latitude",
          "longitude"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "facilityType": {
            "type": "integer",
            "format": "int32"
          },
          "isAccessible": {
            "type": "boolean"
          },
          "latitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateFacilityRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "facilityType": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isAccessible": {
            "type": "boolean",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "FacilityImageResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "imageUrl"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "altText": {
            "type": "string",
            "nullable": true
          },
          "caption": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "OutgoingFriendRequestResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "receiverIdentifier",
          "receiverNickName"
        ],
        "properties": {
          "receiverIdentifier": {
            "type": "string"
          },
          "receiverNickName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SendFriendRequestRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "receiverNickName"
        ],
        "properties": {
          "receiverNickName": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Skriv: Bearer {din Keycloak access token}",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}